requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
aiohttp>=3.9
pandas>=2.1
numpy>=1.26
python-dotenv>=1.0
APScheduler>=3.10
